Refactor database reset handling in AutoRunEngine and Websocket hooked browser specs for improved synchronization

+++ spec/beef/core/main/autorun_engine/autorun_engine_spec.rb
@@ -10,6 +10,8 @@ require_relative '../../../../spec_helper'
 require_relative '../../../../support/constants'
 require_relative '../../../../support/beef_test'
 
+MUTEX = Mutex.new
+
 RSpec.describe 'AutoRunEngine Test', run_on_browserstack: true do
   before(:all) do
     @config = BeEF::Core::Configuration.instance
@@ -18,7 +20,10 @@ RSpec.describe 'AutoRunEngine Test', run_on_browserstack: true do
     print_info 'Loading database'
     db_file = @config.get('beef.database.file')
     print_info 'Resetting the database for BeEF.'
-    File.delete(db_file) if File.exist?(db_file)
+
+    if ENV['RESET_DB']
+      File.delete(db_file) if File.exist?(db_file)
+    end
 
     @config.set('beef.credentials.user', 'beef')
     @config.set('beef.credentials.passwd', 'beef')
@@ -49,9 +54,13 @@ RSpec.describe 'AutoRunEngine Test', run_on_browserstack: true do
       OTR::ActiveRecord.establish_connection!
     end
     ActiveRecord::Migrator.migrations_paths = [File.join('core', 'main', 'ar-migrations')]
-    context = ActiveRecord::MigrationContext.new(ActiveRecord::Migrator.migrations_paths)
-    ActiveRecord::Migrator.new(:up, context.migrations, context.schema_migration, context.internal_metadata).migrate if context.needs_migration?
-
+    MUTEX.synchronize do
+      context = ActiveRecord::MigrationContext.new(ActiveRecord::Migrator.migrations_paths)
+      if context.needs_migration?
+        ActiveRecord::Migrator.new(:up, context.migrations, context.schema_migration, context.internal_metadata).migrate
+      end
+    end
+    
     BeEF::Core::Migration.instance.update_db!
 
     # add AutoRunEngine rule
